# services/venue-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  venue-service_internal_dto.AddAmenityRequest:
    properties:
      amenity_id:
        type: integer
    required:
    - amenity_id
    type: object
...
  venue-service_internal_dto.CreateVenueRequest:
    properties:
      address:
        type: string
      city:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - address
    - name
    type: object
...
  venue-service_internal_dto.UpdateVenueRequest:
    properties:
      address:
        type: string
      city:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - address
    - name
    type: object
host: localhost:8083
info:
  contact: {}
  description: This is the venue service API for the coworking booking system.
  title: Venue Service API
  version: "1.0"
paths:
...
  /admin/venues:
    get:
      description: List venues by status (Admin/Moderator only)
      parameters:
      - description: 'Status filter: pending / approved / blocked'
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List venues with filter
      tags:
      - Admin Venue
  /admin/venues/{id}/approve:
    put:
      description: Approve a venue by ID (Admin/Moderator only)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Approve a venue
      tags:
      - Admin Venue
  /admin/venues/{id}/block:
    put:
      description: Block a venue by ID (Admin/Moderator only)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Block a venue
      tags:
      - Admin Venue
...
  /venues:
    get:
      description: Get list of venues accessible by the user
      parameters:
      - description: Filter by city
        in: query
        name: city
        type: string
      - description: Filter by name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all venues
      tags:
      - Venue
    post:
      consumes:
      - application/json
      description: Create a new venue (user must be authenticated)
      parameters:
      - description: CreateVenueRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/venue-service_internal_dto.CreateVenueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new venue
      tags:
      - Venue
  /venues/{id}:
    delete:
      description: Delete a venue by ID (user must be authenticated)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a venue
      tags:
      - Venue
    get:
      description: Get detailed info of a venue by ID
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get venue by ID
      tags:
      - Venue
    put:
      consumes:
      - application/json
      description: Update a venue by ID (user must be authenticated)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateVenueRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/venue-service_internal_dto.UpdateVenueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a venue
      tags:
      - Venue
  /venues/{id}/amenities:
    post:
      consumes:
      - application/json
      description: Add an existing amenity to a venue (user must be authenticated)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      - description: AddAmenityRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/venue-service_internal_dto.AddAmenityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add an amenity to a venue
      tags:
      - Venue Amenity
  /venues/{id}/amenities/{venueAmenityId}:
    delete:
      description: Remove a specific amenity from a venue (user must be authenticated)
      parameters:
      - description: Venue Amenity ID
        in: path
        name: venueAmenityId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove an amenity from a venue
      tags:
      - Venue Amenity
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"